Moved MCP utils into its own module
